+++ core/lib/template_provider.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const crypto = require('crypto');
 const fs = require('fs');
 const path = require('path');
 
@@ -7,48 +8,173 @@ const ResourceCache = require('./resource_cache').ResourceCache;
 const Template = require('./template').Template;
 const { FsSchemaProvider } = require('./schema_provider');
 
-class FsTemplateProvider {
-    constructor(templateRootPath, filteredSets) {
-        this.config_template_path = templateRootPath;
-        this.schemaProviders = {};
-        this.filteredSets = new Set(filteredSets || []);
+// Known good hashes for template sets
+const supportedHashes = {
+    'bigip-fast-templates': [
+        '47b595bb0bd4f471fe6f057ca1f1e71f5dc84ab2be420ae2a555aca7ea311989'
+    ],
+    examples: [
+        '01e426ecd5457c46c0a59eac767b0fcb13085dc0a661b76f542a0b4b051fff60'
+    ]
+};
 
-        this.cache = new ResourceCache((templateName) => {
-            const tsName = templateName.split('/')[0];
-            if (!this.schemaProviders[tsName]) {
-                this.schemaProviders[tsName] = new FsSchemaProvider(`${templateRootPath}/${tsName}`);
-            }
-            const schemaProvider = this.schemaProviders[tsName];
-            let useMst = 0;
-            let tmplpath = `${templateRootPath}/${templateName}`;
-            if (fs.existsSync(`${tmplpath}.yml`)) {
-                tmplpath = `${tmplpath}.yml`;
-            } else if (fs.existsSync(`${tmplpath}.yaml`)) {
-                tmplpath = `${tmplpath}.yaml`;
-            } else if (fs.existsSync(`${tmplpath}.mst`)) {
-                useMst = 1;
-                tmplpath = `${tmplpath}.mst`;
-            } else {
-                return Promise.reject(new Error(`could not find a template with name "${templateName}"`));
-            }
 
-            return new Promise((resolve, reject) => {
-                fs.readFile(tmplpath, (err, data) => {
-                    if (err) reject(err);
-                    else {
-                        resolve(data.toString('utf8'));
-                    }
-                });
-            }).then(tmpldata => Template[(useMst) ? 'loadMst' : 'loadYaml'](tmpldata, schemaProvider));
+class BaseTemplateProvider {
+    constructor() {
+        if (new.target === BaseTemplateProvider) {
+            throw new TypeError('Cannot instantiate Abstract BaseTemplateProvider');
+        }
+
+        const abstractMethods = [
+            '_loadTemplate',
+            'listSets',
+            'removeSet',
+            'list',
+            'getSchemas'
+        ];
+        abstractMethods.forEach((method) => {
+            if (this[method] === undefined) {
+                throw new TypeError(`Expected ${method} to be defined`);
+            }
         });
+
+
+        this.cache = new ResourceCache((tmplName => this._loadTemplate(tmplName)));
+    }
+
+    invalidateCache() {
+        this.cache.invalidate();
     }
 
     fetch(key) {
         return this.cache.fetch(key);
     }
 
-    invalidateCache() {
-        this.cache.invalidate();
+    fetchSet(setName) {
+        return this.list(setName)
+            .then(tmplList => Promise.all(tmplList.map(tmplName => Promise.all([
+                Promise.resolve(tmplName),
+                this.fetch(tmplName)
+            ]))))
+            .then(tmplList => tmplList.reduce((acc, curr) => {
+                const [tmplName, tmplData] = curr;
+                acc[tmplName] = tmplData;
+                return acc;
+            }, {}));
+    }
+
+    hasSet(setid) {
+        return this.listSets()
+            .then(sets => sets.includes(setid));
+    }
+
+    getNumTemplateSourceTypes(filteredSetName) {
+        const sourceTypes = {};
+        const filteredSetList = (filteredSetName && [filteredSetName]) || [];
+        return this.list(filteredSetList)
+            .then(tmplList => Promise.all(tmplList.map(tmpl => this.fetch(tmpl))))
+            .then(tmplList => tmplList.forEach((tmpl) => {
+                if (!sourceTypes[tmpl.sourceType]) {
+                    sourceTypes[tmpl.sourceType] = 0;
+                }
+                sourceTypes[tmpl.sourceType] += 1;
+            }))
+            .then(() => sourceTypes);
+    }
+
+    getNumSchema(filteredSetName) {
+        return this.getSchemas(filteredSetName)
+            .then(schemas => Object.keys(schemas).length);
+    }
+
+    getSetData(setName) {
+        return Promise.all([
+            this.fetchSet(setName),
+            this.getSchemas(setName)
+        ])
+            .then(([templates, schemas]) => {
+                if (Object.keys(templates).length === 0) {
+                    return Promise.reject(new Error(`No templates found for template set ${setName}`));
+                }
+                const tsHash = crypto.createHash('sha256');
+                const tmplHashes = Object.values(templates).map(x => x.sourceHash).sort();
+                tmplHashes.forEach((hash) => {
+                    tsHash.update(hash);
+                });
+                const schemaHashes = Object.values(schemas).map(x => x.hash).sort();
+                schemaHashes.forEach((hash) => {
+                    tsHash.update(hash);
+                });
+
+                const tsHashDigest = tsHash.digest('hex');
+                const supported = (
+                    Object.keys(supportedHashes).includes(setName)
+                    && supportedHashes[setName].includes(tsHashDigest)
+                );
+                return Promise.resolve({
+                    name: setName,
+                    hash: tsHashDigest,
+                    supported,
+                    templates: Object.keys(templates).reduce((acc, curr) => {
+                        const tmpl = templates[curr];
+                        acc.push({
+                            name: curr,
+                            hash: tmpl.sourceHash
+                        });
+                        return acc;
+                    }, []),
+                    schemas: Object.keys(schemas).reduce((acc, curr) => {
+                        const schema = schemas[curr];
+                        acc.push({
+                            name: schema.name,
+                            hash: schema.hash
+                        });
+                        return acc;
+                    }, [])
+                });
+            });
+    }
+}
+
+class FsTemplateProvider extends BaseTemplateProvider {
+    constructor(templateRootPath, filteredSets) {
+        super();
+        this.config_template_path = templateRootPath;
+        this.schemaProviders = {};
+        this.filteredSets = new Set(filteredSets || []);
+    }
+
+    _loadTemplate(templateName) {
+        const tsName = templateName.split('/')[0];
+        this._ensureSchemaaProvider(tsName);
+        const schemaProvider = this.schemaProviders[tsName];
+        let useMst = 0;
+        let tmplpath = `${this.config_template_path}/${templateName}`;
+        if (fs.existsSync(`${tmplpath}.yml`)) {
+            tmplpath = `${tmplpath}.yml`;
+        } else if (fs.existsSync(`${tmplpath}.yaml`)) {
+            tmplpath = `${tmplpath}.yaml`;
+        } else if (fs.existsSync(`${tmplpath}.mst`)) {
+            useMst = 1;
+            tmplpath = `${tmplpath}.mst`;
+        } else {
+            return Promise.reject(new Error(`could not find a template with name "${templateName}"`));
+        }
+
+        return new Promise((resolve, reject) => {
+            fs.readFile(tmplpath, (err, data) => {
+                if (err) reject(err);
+                else {
+                    resolve(data.toString('utf8'));
+                }
+            });
+        }).then(tmpldata => Template[(useMst) ? 'loadMst' : 'loadYaml'](tmpldata, schemaProvider));
+    }
+
+    _ensureSchemaaProvider(tsName) {
+        if (!this.schemaProviders[tsName]) {
+            this.schemaProviders[tsName] = new FsSchemaProvider(`${this.config_template_path}/${tsName}`);
+        }
     }
 
     listSets() {
@@ -85,24 +211,15 @@ class FsTemplateProvider {
                             })
                     );
                 });
-            })))).then(sets => sets.reduce((acc, curr) => acc.concat(curr), []));
+            })))).then(sets => sets.reduce((acc, curr) => acc.concat(curr), []))
+            .then((sets) => {
+                sets.forEach(set => this._ensureSchemaaProvider(set));
+                return sets;
+            });
     }
 
-    getNumTemplateSourceTypes(filteredSetName) {
-        const sourceTypes = {};
-        const filteredSetList = (filteredSetName && [filteredSetName]) || [];
-        return this.list(filteredSetList)
-            .then(tmplList => Promise.all(tmplList.map(tmpl => this.fetch(tmpl))))
-            .then(tmplList => tmplList.forEach((tmpl) => {
-                if (!sourceTypes[tmpl.sourceType]) {
-                    sourceTypes[tmpl.sourceType] = 0;
-                }
-                sourceTypes[tmpl.sourceType] += 1;
-            }))
-            .then(() => sourceTypes);
-    }
-
-    getNumSchema(filteredSetName) {
+    getSchemas(filteredSetName) {
+        const schemas = {};
         return Promise.resolve()
             .then(() => {
                 if (filteredSetName) {
@@ -110,13 +227,27 @@ class FsTemplateProvider {
                 }
                 return this.listSets();
             })
-            .then(setList => setList.map(x => this.schemaProviders[x].list()))
-            .then(schemaLists => schemaLists.flat().length);
-    }
-
-    hasSet(setid) {
-        return this.listSets()
-            .then(sets => sets.includes(setid));
+            .then((setList) => {
+                setList.forEach(tsName => this._ensureSchemaaProvider(tsName));
+                return setList;
+            })
+            .then(setList => Promise.all(setList.map(
+                tsName => this.schemaProviders[tsName].list()
+                    .then(schemaList => Promise.all(schemaList.map(
+                        schemaName => this.schemaProviders[tsName].fetch(schemaName)
+                            .then((schemaData) => {
+                                const name = `${tsName}/${schemaName}`;
+                                const schemaHash = crypto.createHash('sha256');
+                                schemaHash.update(schemaData);
+                                schemas[name] = {
+                                    name,
+                                    data: schemaData,
+                                    hash: schemaHash.digest('hex')
+                                };
+                            })
+                    )))
+            )))
+            .then(() => schemas);
     }
 
     removeSet() {
@@ -124,40 +255,37 @@ class FsTemplateProvider {
     }
 }
 
-class DataStoreTemplateProvider {
+class DataStoreTemplateProvider extends BaseTemplateProvider {
     constructor(datastore, filteredSets) {
+        super();
         this.filteredSets = new Set(filteredSets || []);
         this.storage = datastore;
         this.keyCache = [];
         this._numSchema = {};
-
-        this.cache = new ResourceCache((templatePath) => {
-            const [tsName, templateName] = templatePath.split('/');
-            return this.storage.hasItem(tsName)
-                .then((result) => {
-                    if (result) {
-                        return Promise.resolve();
-                    }
-                    return Promise.reject(new Error(`Could not find template set "${tsName}" in data store`));
-                })
-                .then(() => this.storage.getItem(tsName))
-                .then((tsData) => {
-                    const templateData = tsData.templates[templateName];
-
-                    if (typeof templateData === 'undefined') {
-                        return Promise.reject(new Error(`Could not find template "${templateName}" in template set "${tsName}"`));
-                    }
-                    return Template.fromJson(JSON.parse(templateData));
-                });
-        });
     }
 
-    fetch(key) {
-        return this.cache.fetch(key);
+    _loadTemplate(templatePath) {
+        const [tsName, templateName] = templatePath.split('/');
+        return this.storage.hasItem(tsName)
+            .then((result) => {
+                if (result) {
+                    return Promise.resolve();
+                }
+                return Promise.reject(new Error(`Could not find template set "${tsName}" in data store`));
+            })
+            .then(() => this.storage.getItem(tsName))
+            .then((tsData) => {
+                const templateData = tsData.templates[templateName];
+
+                if (typeof templateData === 'undefined') {
+                    return Promise.reject(new Error(`Could not find template "${templateName}" in template set "${tsName}"`));
+                }
+                return Template.fromJson(JSON.parse(templateData));
+            });
     }
 
     invalidateCache() {
-        this.cache.invalidate();
+        super.invalidateCache();
         this.keyCache = [];
     }
 
@@ -191,11 +319,6 @@ class DataStoreTemplateProvider {
             });
     }
 
-    hasSet(setid) {
-        return this.listSets()
-            .then(sets => sets.includes(setid));
-    }
-
     removeSet(setid) {
         return this.hasSet(setid)
             .then(result => (result || Promise.reject(
@@ -205,21 +328,7 @@ class DataStoreTemplateProvider {
             .then(() => this.invalidateCache());
     }
 
-    getNumTemplateSourceTypes(filteredSetName) {
-        const sourceTypes = {};
-        const filteredSetList = (filteredSetName && [filteredSetName]) || [];
-        return this.list(filteredSetList)
-            .then(tmplList => Promise.all(tmplList.map(tmpl => this.fetch(tmpl))))
-            .then(tmplList => tmplList.forEach((tmpl) => {
-                if (!sourceTypes[tmpl.sourceType]) {
-                    sourceTypes[tmpl.sourceType] = 0;
-                }
-                sourceTypes[tmpl.sourceType] += 1;
-            }))
-            .then(() => sourceTypes);
-    }
-
-    getNumSchema(filteredSetName) {
+    getSchemas(filteredSetName) {
         return Promise.resolve()
             .then(() => {
                 if (filteredSetName) {
@@ -228,7 +337,22 @@ class DataStoreTemplateProvider {
                 return this.listSets();
             })
             .then(setNames => Promise.all(setNames.map(x => this.storage.getItem(x))))
-            .then(sets => sets.reduce((acc, curr) => acc + Object.keys(curr.schemas).length, 0));
+            .then(setData => setData.filter(x => x))
+            .then(setData => setData.reduce((acc, curr) => {
+                const tsName = curr.name;
+                Object.keys(curr.schemas).forEach((schemaName) => {
+                    const schemaData = curr.schemas[schemaName];
+                    const name = `${tsName}/${schemaName}`;
+                    const schemaHash = crypto.createHash('sha256');
+                    schemaHash.update(schemaData);
+                    acc[name] = {
+                        name,
+                        data: schemaData,
+                        hash: schemaHash.digest('hex')
+                    };
+                });
+                return acc;
+            }, {}));
     }
 
     static fromFs(datastore, templateRootPath, filteredSets) {