+++ iappslx/nodejs/templateWorker.js
@@ -3,7 +3,6 @@
 require('core-js');
 
 const fs = require('fs-extra');
-const crypto = require('crypto');
 
 const yaml = require('js-yaml');
 const extract = require('extract-zip');
@@ -29,16 +28,6 @@ const scratchPath = `${configPath}/scratch`;
 const uploadPath = '/var/config/rest/downloads';
 const dataGroupPath = `/Common/${projectName}/dataStore`;
 
-// Known good hashes for template sets. Always keep the latest at index 0!
-const supportedHashes = {
-    'bigip-fast-templates': [
-        'b29e5ebeb19803cf382bea0a033f1351d374ca327386ff6102deb44721caf2cb'
-    ],
-    examples: [
-        'ecf1dce5b54ecab68567c4e26a7bbee777f8f6b2affd005d408f4699192ffb1b'
-    ]
-};
-
 class TemplateWorker {
     constructor() {
         this.state = {};
@@ -55,10 +44,21 @@ class TemplateWorker {
         });
     }
 
+    hookCompleteRestOp() {
+        // Hook completeRestOperation() so we can add additional logging
+        this._prevCompleteRestOp = this.completeRestOperation;
+        this.completeRestOperation = (restOperation) => {
+            this.recordRestResponse(restOperation);
+            this._prevCompleteRestOp(restOperation);
+        };
+    }
+
     /**
      * Worker Handlers
      */
     onStart(success, error) {
+        this.hookCompleteRestOp();
+
         // Find any template sets on disk (e.g., from the RPM) and add them to
         // the data store. Do not overwrite template sets already in the data store.
         const fsprovider = new FsTemplateProvider(templatesPath);
@@ -186,7 +186,10 @@ class TemplateWorker {
     generateTeemReportError(restOp) {
         const uri = restOp.getUri();
         const pathElements = uri.path.split('/');
-        const endpoint = pathElements.slice(0, 4).join('/');
+        let endpoint = pathElements.slice(0, 4).join('/');
+        if (pathElements[4]) {
+            endpoint = `${endpoint}/item`;
+        }
         const report = {
             method: restOp.getMethod(),
             endpoint,
@@ -200,50 +203,19 @@ class TemplateWorker {
     /**
      * Helper functions
      */
-    gatherTemplates(setList) {
-        return this.templateProvider.list(setList)
-            .then(tmplList => Promise.all(tmplList.map(tmplName => Promise.all([
-                Promise.resolve(tmplName),
-                this.templateProvider.fetch(tmplName)
-            ]))))
-            .then(tmplList => tmplList.reduce((acc, curr) => {
-                const [tmplName, tmplData] = curr;
-                acc[tmplName] = tmplData;
-                return acc;
-            }, {}));
-    }
-
     gatherTemplateSet(tsid) {
-        return this.gatherTemplates(tsid)
-            .then((templates) => {
-                if (Object.keys(templates).length === 0) {
-                    return Promise.reject(new Error(`No templates found for template set ${tsid}`));
-                }
-                const tsHash = crypto.createHash('sha256');
-                const tmplHashes = Object.values(templates).map(x => x.sourceHash).sort();
-                tmplHashes.forEach((hash) => {
-                    tsHash.update(hash);
-                });
-                const tsHashDigest = tsHash.digest('hex');
-                const supported = (
-                    Object.keys(supportedHashes).includes(tsid)
-                    && supportedHashes[tsid].includes(tsHashDigest)
+        const fsprovider = new FsTemplateProvider(templatesPath);
+        return Promise.all([
+            this.templateProvider.getSetData(tsid),
+            fsprovider.hasSet(tsid)
+                .then(result => (result ? fsprovider.getSetData(tsid) : Promise.resolve(undefined)))
+        ])
+            .then(([tsData, fsTsData]) => {
+                tsData.updateAvailable = (
+                    fs.existsSync(`${templatesPath}/${tsid}`)
+                    && fsTsData && fsTsData.hash !== tsData.hash
                 );
-                const updateAvailable = fs.existsSync(`${templatesPath}/${tsid}`);
-                return Promise.resolve({
-                    name: tsid,
-                    hash: tsHashDigest,
-                    supported,
-                    updateAvailable,
-                    templates: Object.keys(templates).reduce((acc, curr) => {
-                        const tmpl = templates[curr];
-                        acc.push({
-                            name: curr,
-                            hash: tmpl.sourceHash
-                        });
-                        return acc;
-                    }, [])
-                });
+                return tsData;
             });
     }
 
@@ -272,6 +244,27 @@ class TemplateWorker {
     /**
      * HTTP/REST handlers
      */
+    recordRestRequest(restOp) {
+        this.logger.fine(
+            `TemplateWorker received request: method=${restOp.getMethod()}; path=${restOp.getUri().path}; data=${JSON.stringify(restOp.body)}`
+        );
+    }
+
+    recordRestResponse(restOp) {
+        const minOp = {
+            method: restOp.getMethod(),
+            path: restOp.getUri().path,
+            status: restOp.getStatusCode(),
+            body: restOp.getBody()
+        };
+        const msg = `TemplateWorker sending response:\n${JSON.stringify(minOp, null, 2)}`;
+        if (minOp.status >= 400) {
+            this.logger.error(msg);
+        } else {
+            this.logger.fine(msg);
+        }
+    }
+
     genRestResponse(restOperation, code, message) {
         restOperation.setStatusCode(code);
         restOperation.setBody({
@@ -391,6 +384,9 @@ class TemplateWorker {
         const pathElements = uri.path.split('/');
         const collection = pathElements[3];
         const itemid = pathElements[4];
+
+        this.recordRestRequest(restOperation);
+
         try {
             switch (collection) {
             case 'info':
@@ -456,12 +452,13 @@ class TemplateWorker {
         const tsid = data.name;
         const setpath = `${uploadPath}/${tsid}.zip`;
         const scratch = `${scratchPath}/${tsid}`;
+        const onDiskPath = `${templatesPath}/${tsid}`;
 
         if (!data.name) {
             return this.genRestResponse(restOperation, 400, `invalid template set name supplied: ${tsid}`);
         }
 
-        if (!fs.existsSync(setpath) && !Object.keys(supportedHashes).includes(tsid)) {
+        if (!fs.existsSync(setpath) && !fs.existsSync(onDiskPath)) {
             return this.genRestResponse(restOperation, 404, `${setpath} does not exist`);
         }
 
@@ -471,8 +468,8 @@ class TemplateWorker {
 
         return Promise.resolve()
             .then(() => {
-                if (Object.keys(supportedHashes).includes(tsid)) {
-                    return fs.copy(`${templatesPath}/${tsid}`, scratch);
+                if (fs.existsSync(onDiskPath)) {
+                    return fs.copy(onDiskPath, scratch);
                 }
 
                 return new Promise((resolve, reject) => {
@@ -500,6 +497,8 @@ class TemplateWorker {
         const uri = restOperation.getUri();
         const pathElements = uri.path.split('/');
         const collection = pathElements[3];
+
+        this.recordRestRequest(restOperation);
 
         try {
             switch (collection) {
@@ -562,6 +561,8 @@ class TemplateWorker {
         const pathElements = uri.path.split('/');
         const collection = pathElements[3];
         const itemid = pathElements[4];
+
+        this.recordRestRequest(restOperation);
 
         try {
             switch (collection) {